Split up AdminController and create AdminProductCategoryController
